+++ src/LuaPlugins/LuaPluginAPI.hpp
@@ -12,13 +12,13 @@ struct LuaServer {
 	static void LuaSystemWideMessage(std::string message);
 	static void LuaBroadcastMessage(std::string message);
 	static void LuaKickClient(Client* client, std::string reason);
-	static void LuaSendBlock(Client* client, short x, short y, short z, int type);
+	static void LuaSendBlock(Client* client, short x, short y, short z, uint8_t type);
 	static Client* LuaGetClientByName(std::string name, bool exact);
 	static void LuaLoadPlugin(std::string filename);
 	static void LuaRegisterEvent(int type, luabridge::LuaRef func);
 	static LuaCommand* LuaAddCommand(std::string name, std::string aliases, luabridge::LuaRef func, std::string docString,
 		unsigned argumentAmount, unsigned permissionLevel);
-	static void LuaPlaceBlock(Client* client, int type, short x, short y, short z);
+	static void LuaPlaceBlock(Client* client, uint8_t type, short x, short y, short z);
 	static int LuaMapGetBlockType(Client* client, short x, short y, short z);
 	static void LuaSendKick(Client* client, std::string reason);
 	static luabridge::LuaRef LuaGetClients();
